+++ packages/collections-domain/src/CriteriaRecollection.php
@@ -22,7 +22,6 @@ use Rekalogika\Contracts\Rekapager\PageableInterface;
 use Rekalogika\Domain\Collections\Common\CountStrategy;
 use Rekalogika\Domain\Collections\Common\Trait\ItemsWithSafeguardTrait;
 use Rekalogika\Domain\Collections\Common\Trait\PageableTrait;
-use Rekalogika\Domain\Collections\Common\Trait\ReadableCollectionTrait;
 use Rekalogika\Domain\Collections\Exception\UnexpectedValueException;
 use Rekalogika\Domain\Collections\Trait\ReadableExtraLazyTrait;
 use Rekalogika\Domain\Collections\Trait\RecollectionTrait;
@@ -37,9 +36,6 @@ class CriteriaRecollection implements PageableInterface, ReadableCollection
 {
     /** @use RecollectionTrait<TKey,T> */
     use RecollectionTrait;
-
-    /** @use ReadableCollectionTrait<TKey,T> */
-    use ReadableCollectionTrait;
 
     /** @use PageableTrait<TKey,T> */
     use PageableTrait;