+++ packages/collections-domain/src/Trait/ExtraLazyTrait.php
@@ -13,6 +13,11 @@ declare(strict_types=1);
 
 namespace Rekalogika\Domain\Collections\Trait;
 
+use Rekalogika\Domain\Collections\Common\Trait\ArrayAccessTrait;
+use Rekalogika\Domain\Collections\Common\Trait\CountableTrait;
+use Rekalogika\Domain\Collections\Common\Trait\IteratorAggregateTrait;
+use Rekalogika\Domain\Collections\Common\Trait\WritableCollectionTrait;
+
 /**
  * @template TKey of array-key
  * @template T
@@ -21,8 +26,23 @@ namespace Rekalogika\Domain\Collections\Trait;
  */
 trait ExtraLazyTrait
 {
-    /** @use ReadableExtraLazyTrait<TKey,T> */
-    use ReadableExtraLazyTrait;
+    /**
+     * @use WritableCollectionTrait<TKey,T>
+     * @use ReadableExtraLazyTrait<TKey,T>
+     */
+    use WritableCollectionTrait, ReadableExtraLazyTrait {
+        WritableCollectionTrait::filter insteadof ReadableExtraLazyTrait;
+        WritableCollectionTrait::map insteadof ReadableExtraLazyTrait;
+        WritableCollectionTrait::partition insteadof ReadableExtraLazyTrait;
+    }
+
+    use CountableTrait;
+
+    /** @use IteratorAggregateTrait<TKey,T> */
+    use IteratorAggregateTrait;
+
+    /** @use ArrayAccessTrait<TKey,T> */
+    use ArrayAccessTrait;
 
     //
     // ArrayAccess