+++ packages/collections-domain/src/Trait/RecollectionTrait.php
@@ -60,4 +60,18 @@ trait RecollectionTrait
 
         return $this->pageable;
     }
+
+    /**
+     * @return int<0,max>
+     */
+    private function getRealCount(): int
+    {
+        $count = $this->collection->count();
+
+        if ($count > 0) {
+            return $count;
+        }
+
+        return 0;
+    }
 }